CRUD

+++ index.php
@@ -9,55 +9,16 @@ if (!$userList) {
 }
 
 if (isset($_GET['action']) && isset($_GET['id'])) {
-    $id = $_GET['id'];
-    $req = "DELETE FROM utilisateurs WHERE id = $id";
-    mysqli_query($connexion, $req);
-    header("Location: index.php");
+    if ($_GET['action'] == 'delete') {
+        $id = $_GET['id'];
+        $req = "DELETE FROM utilisateurs WHERE id = $id";
+        mysqli_query($connexion, $req);
+        header("Location: index.php");
+    } elseif ($_GET['action'] == 'modif') {
+        $id = $_GET['id'];
+        header("Location: modif.php?id=$id");
+    }
 }
 
-?>
-<!DOCTYPE html>
-<html lang="en">
 
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-LN+7fdVzj6u52u30Kp6M/trliBMCMKTyK833zpbD+pXdCLuTusPj697FH4R/5mcr" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/js/bootstrap.bundle.min.js" integrity="sha384-ndDqU0Gzau9qJ1lfW4pNLlhNTkCfHzAVBReH9diLvGRem5+R9g2FzA8ZGN954O5Q" crossorigin="anonymous"></script>
-</head>
-
-<body>
-    <div class="container">
-        <h1>Liste des utilisateurs</h1>
-        <a href="ajout.php" class="btn btn-4 btn-primary float-end">Ajouter</a>
-
-        <table class="table table-striped">
-            <thead>
-                <tr>
-                    <th>ID</th>
-                    <th>Nom</th>
-                    <th>Prenom</th>
-                    <th>Email</th>
-                    <th>Actions</th>
-                </tr>
-            </thead>
-            <tbody>
-                <?php while ($user = mysqli_fetch_assoc($userList)) { ?>
-                    <tr>
-                        <td><?= $user['id'] ?></td>
-                        <td><?= $user['nom'] ?></td>
-                        <td><?= $user['prenom'] ?></td>
-                        <td><?= $user['email'] ?></td>
-                        <td>
-                            <a class="btn btn-4 btn-warning">Modifier</a>
-                            <a href="?action=delete&id=<?= $user['id'] ?>" class="btn btn-4 btn-danger">Supprimer</a>
-                        </td>
-                    </tr>
-                <?php } ?>
-            </tbody>
-        </table>
-    </div>
-</body>
-
-</html>
+require_once 'liste.php';
